Refactoring of previous code
The current code does not yet work!

Basic design is described below:
The idea is that I process all the lines in a GCode script in an unsequenced parallel way (the logic for this is basically the same in C++17 as processing this in sequence one 1 thread) I only need to sort it afterwards (but that is simple).
Each line is split up in a string_view (pointer to the first char and the length, non ownership). Using the ranges-v3 library I split each line in a comment section and instruction setting (neglecting for now that comments might also contain instructions for the printer). These are also string_views.
The instruction section is split as string_views at each space character also as a ranges-view, where the first sub-range-view is regarded as the cmd (G0, G1, M104) and the other subranges are a variable number of subranges with a parameter (E, X, Y, Z etc.) stored as string_view and value where the text is stored as double after a conversion with stod
These processed instructions are stored in a data type called.... tatatata Instruction
Which also include a line_number.
The vector of processed instructions is sorted by line number.
The state space vector of the printer is created, this vector contains all the variables at a given time. The first state space vector describes the starting state of the print (x=0mm, y=0mm, z=0mm, e=0mm, t=0s, T_1=273.15K etc.).
This state space vector is then transformed when all instructions are processed in a range based loop. Each Instruction instance has a (state_space_vector_previous) operator which returns a new state_space_vector with the instructions. In this ()
operator the required function is obtained from a class where two maps are stored const map<string_view, function_ptr>
 the previous obtained string_view with the command is passed to a getter, which returns the required transform function. A same kind of mapping principle will be used for the parameters. These transformations are done in sequence.
The benefit of this way is that it allows me to separate logistical logic  from GCode logic, and it would also allow for different dialects of GCode and specific printers with not standard state_space_vectors.
The vector with transformed state_space_vectors are then converted to a Numpy array or passed along with protobuf to either Cura for simulation or a Python Pandas script for statistical analysis.