Commands are now vectorized extracted
A new optional instruction vector is created, which serves as a place to store the optional extracted
commands. This vector is filled with a transform applied over the lines. A second step all lines that
weren't converted to an instruction (nullptr) are removed from this vector. Finally the remaining
optional Instructions are used to transform the cmdlines vector with Instructions. This vector
is then sorted.

CURA-7525